fix(exo-shell): localize the Spawner type in the public .d.ts (no devDep leak)
Panel round-2 (typist) should-fix: the shipped `declare module '@endo/exo-shell'`
typed makeShell's spawner as `import('@endo/host-spawner').Spawner`, but exo-shell
lists @endo/host-spawner only as a devDependency. In-repo tsc is clean, but an
external consumer installing @endo/exo-shell without host-spawner would get
`Cannot find module` and spawner would silently degrade to `any`. Declare a local
structural `Spawner` type instead (matching how this package already localizes
ShellPolicy/ShellResult/EndoShell); the host `makeHostSpawner`'s Spawner remains
assignable to it. The implementation's JSDoc keeps importing the host-spawner
type in-repo, which is fine for a devDependency.
