+++ packages/exo-shell/package.json
@@ -20,19 +20,23 @@
   },
   "type": "module",
   "main": "./src/index.js",
-  "types": "./types.d.ts",
+  "types": "./types-index.d.ts",
   "exports": {
     ".": {
-      "types": "./types.d.ts",
+      "types": "./types-index.d.ts",
       "default": "./src/index.js"
     },
     "./src/shell.js": {
-      "types": "./types.d.ts",
+      "types": "./src/shell.d.ts",
       "default": "./src/shell.js"
     },
     "./src/interfaces.js": {
-      "types": "./types.d.ts",
+      "types": "./src/interfaces.d.ts",
       "default": "./src/interfaces.js"
+    },
+    "./types-index.js": {
+      "types": "./types-index.d.ts",
+      "default": "./types-index.js"
     },
     "./package.json": "./package.json"
   },
