+++ packages/exo-shell/types-index.d.ts
@@ -1,0 +1,2 @@
+export type * from './src/types.js';
+export { makeShell, ShellInterface } from './src/index.js';
