+++ packages/exo-shell/types-index.js
@@ -1,0 +1,2 @@
+export { makeShell } from './src/shell.js';
+export { ShellInterface } from './src/interfaces.js';
