+++ packages/host-spawner/types-index.d.ts
@@ -1,0 +1,2 @@
+export type * from './src/types.js';
+export { makeHostSpawner } from './src/index.js';
