+++ packages/host-spawner/types-index.js
@@ -1,0 +1,1 @@
+export { makeHostSpawner } from './src/host-spawner.js';
